Move cache implementation to folder
